requirements.txt
django>=5.0
telethon
python-dotenv
asyncpg
gunicorn
psycopg2-binary
whitenoise
